perf: type-annotate the status-dictionary reads
`status` is a `Dict{Symbol, Any}`, so a bare `n = m.status[:N]` infers to `Any`.
That is not confined to `n`: the range, the views taken from it, and the
accumulator all become dynamic, so the whole enclosing function boxes on every
inner-loop iteration. The keys read this way (`:N`, `:d`, `:d_unique`,
`:d⊥_unique`) are all `length(...)`, so annotating them `::Int` restores
inference. All 52 sites, no other change.

Worth 50-73x on the CRWCM strength accessors, which were the worst hit because
they call a dyadic probability per inner iteration:

  nonreciprocated_outstrength  41417 ns / 1093 allocs -> 833 ns / 5 allocs
  reciprocated_outstrength     57583 ns / 1783 allocs -> 791 ns / 5 allocs

The DCReM and CReM strength accessors gain a further 2.4-2.7x on top of the
fitness-vector hoist (791 -> 291 ns and 708 -> 291 ns). The UBCM and RBCM
accessors are unaffected, as they do not read the dictionary inside a loop.
Returned values are unchanged, and inference now reports `Float64` where it
previously reported `Any`.

Note for anyone revisiting this: `precision(m) = typeof(m).parameters[2]` looks
like the culprit and is not. It constant-folds and infers to `Type{Float64}`;
rewriting it as a `where {T,N}` clause measures identically. Benchmark behind a
function barrier, since calling through non-const globals boxes the result and
shows a phantom allocation that is not in the code.

+++ src/Models/BiCM.jl
@@ -891,7 +891,7 @@ function solve_model!(m::BiCM;  # common settings
         # on the current AD stack, Zygote's `dict_getindex` pullback errors when a Dict access sits
         # inside the AD path (this is why only the BiCM AD-gradient solve was affected).
         d⊥ᵣ, d⊤ᵣ, f⊥, f⊤, d⊥ᵣ_nz, d⊤ᵣ_nz = m.d⊥ᵣ, m.d⊤ᵣ, m.f⊥, m.f⊤, m.d⊥ᵣ_nz, m.d⊤ᵣ_nz
-        d⊥_unique = m.status[:d⊥_unique]
+        d⊥_unique = m.status[:d⊥_unique]::Int
         f = AD_method ∈ keys(AD_methods)            ? Optimization.OptimizationFunction( (θ, p) ->   -L_BiCM_reduced(θ, d⊥ᵣ, d⊤ᵣ, f⊥, f⊤, d⊥ᵣ_nz, d⊤ᵣ_nz, d⊥_unique),
                                                                                             AD_methods[AD_method],
                                                                                             grad = analytical_gradient ? grad! : nothing)                      : throw(ArgumentError("The AD method $(AD_method) is not supported (yet)"))
